refactor: shared 파일로부터 api 관련 데이터 타입 분리 및 적용
