+++ frontend/src/api/bookmarks.ts
@@ -1,7 +1,7 @@
 import { fetcher } from "@src/api";
 
 import { API_ENDPOINT } from "@src/@constants";
-import { ResponseBookmarks } from "@src/@types/shared";
+import { ResponseBookmarks } from "@src/@types/api";
 
 import { getPrivateHeaders } from "@src/api/utils";
