+++ frontend/src/api/channels.ts
@@ -1,10 +1,7 @@
 import { fetcher } from "@src/api";
 
 import { API_ENDPOINT } from "@src/@constants";
-import {
-  ResponseChannels,
-  ResponseSubscribedChannels,
-} from "@src/@types/shared";
+import { ResponseChannels, ResponseSubscribedChannels } from "@src/@types/api";
 
 import { getPrivateHeaders } from "@src/api/utils";
