+++ frontend/src/api/messages.ts
@@ -1,7 +1,7 @@
 import { fetcher } from "@src/api";
 
 import { API_ENDPOINT } from "@src/@constants";
-import { ResponseMessages } from "@src/@types/shared";
+import { ResponseMessages } from "@src/@types/api";
 
 import { getPrivateHeaders } from "@src/api/utils";
