+++ frontend/src/hooks/@query/useGetInfiniteMessages.ts
@@ -1,7 +1,8 @@
 import { UseInfiniteQueryResult, useInfiniteQuery } from "react-query";
 
 import { QUERY_KEY } from "@src/@constants";
-import { CustomError, ResponseMessages } from "@src/@types/shared";
+import { ResponseMessages } from "@src/@types/api";
+import { CustomError } from "@src/@types/shared";
 
 import { getMessages } from "@src/api/messages";
